+++ src/lib/constants.js
@@ -9,7 +9,7 @@ import path from 'path';
  * @typedef {'none' | 'full'} LightnetProofLevel
  * @typedef {'o1js-main' | 'berkeley' | 'develop'} LightnetMinaBranch
  *
- * @type {{ uiTypes: UiType[], exampleTypes: ExampleType[], feePayerCacheDir: string, lightnetWorkDir: string, lightnetModes: LightnetMode[], lightnetTypes: LightnetType[], lightnetProofLevels: LightnetProofLevel[], lightnetMinaBranches: LightnetMinaBranch[] }}
+ * @type {{ uiTypes: UiType[], exampleTypes: ExampleType[], feePayerCacheDir: string, lightnetWorkDir: string, lightnetModes: LightnetMode[], lightnetTypes: LightnetType[], lightnetProofLevels: LightnetProofLevel[], lightnetMinaBranches: LightnetMinaBranch[], lightnetProcessToLogFileMapping: Map<string, string> }}
  */
 const Constants = Object.freeze({
   uiTypes: ['next', 'svelte', 'nuxt', 'empty', 'none'],
@@ -20,6 +20,18 @@ const Constants = Object.freeze({
   lightnetTypes: ['fast', 'real'],
   lightnetProofLevels: ['none', 'full'],
   lightnetMinaBranches: ['o1js-main', 'berkeley', 'develop'],
+  lightnetProcessToLogFileMapping: new Map([
+    ['Archive-Node-API application', 'logs/archive-node-api.log'],
+    ['Mina Archive process', 'logs/archive-node.log,archive/log.txt'],
+    ['Mina multi-purpose Daemon', 'logs/single-node-network.log'],
+    ['Fish BP #1', 'fish_0/log.txt'],
+    ['Non-consensus node #1', 'node_0/log.txt'],
+    ['Seed node #1', 'seed/log.txt'],
+    ['SNARK coordinator #1', 'snark_coordinator/log.txt'],
+    ['SNARK worker #1', 'snark_workers/worker_0/log.txt'],
+    ['Whale BP #1', 'whale_0/log.txt'],
+    ['Whale BP #2', 'whale_1/log.txt'],
+  ]),
 });
 
 export default Constants;